Move `only_failures` config specs into their own file.
`configuration_spec.rb` is too huge. The API needs to
remain large to support the configurability of RSpec,
but we don’t need to keep all the specs in one file.
I want to start breaking off different chunks into their
own files. This is a first step towards that.
